:recycle: Rewrite JS Code in TS
