+++ src/contracts/authentication/IOidcConfig.ts
@@ -1,0 +1,7 @@
+export interface IOidcConfig {
+  clientId: string;
+  redirectUri: string;
+  logoutRedirectUri: string;
+  responseType: string;
+  scope: string;
+}
