+++ src/contracts/authentication/ITokenObject.ts
@@ -1,4 +1,5 @@
 export interface ITokenObject {
   idToken: string;
   accessToken: string;
+  userId?: string;
 }
