+++ src/contracts/constants.ts
@@ -1,0 +1,6 @@
+export enum StudioVersion {
+  Dev = 'dev',
+  Alpha = 'alpha',
+  Beta = 'beta',
+  Stable = 'stable',
+}
