+++ src/contracts/index.ts
@@ -20,3 +20,4 @@ export * from './inspect-correlation/index';
 export * from './solution-explorer';
 export * from './navigation';
 export * from './diagram/index';
+export * from './constants';
