+++ src/electron_app/oidc-config.ts
@@ -6,8 +6,7 @@
  * Even though this is the redirect route we never truly redirect there since
  * we intercept the redirect request in the implementation.
  */
-
-module.exports = {
+export default {
   clientId: 'bpmn_studio',
   redirectUri: 'http://localhost:9000/signin-oidc',
   logoutRedirectUri: 'http://localhost:9000/signout-oidc',
